move all command-line related stuff to separate file
