+++ Makefile
@@ -16,7 +16,8 @@ endif
 all: csbench
 
 csbench: csbench.c csbench_perf.c csbench_plot.c csbench_utils.c \
-		 csbench_analyze.c csbench_report.c csbench_run.c csbench_serialize.c
+		 csbench_analyze.c csbench_report.c csbench_run.c csbench_serialize.c \
+		 csbench_cli.c
 	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS)
 
 install: csbench
