+++ scripts/amalgamated.py
@@ -4,7 +4,7 @@ import itertools
 
 file_names = ["csbench.h", "csbench.c", "csbench_plot.c", "csbench_perf.c",
               "csbench_utils.c", "csbench_run.c", "csbench_report.c",
-              "csbench_analyze.c", "csbench_serialize.c"]
+              "csbench_analyze.c", "csbench_serialize.c", "csbench_cli.c"]
 files = {}
 for name in file_names:
     with open(name, encoding="utf8") as f:
@@ -26,6 +26,8 @@ def make_core_contents(lines):
 
 amalgamated_source = preample \
         + make_core_contents(files["csbench.h"]) \
+        + ["\n"] \
+        + make_core_contents(files["csbench_cli.c"]) \
         + ["\n"] \
         + make_core_contents(files["csbench_run.c"]) \
         + ["\n"] \
